\name{print.binMto}
\alias{print.binMto}

\title{ print function for objects of class "binMto"}
\description{
 A print functions for objects produced by calling \code{binMto}.
}
\usage{
\method{print}{binMto}(x, digits=4, ...)
}
\arguments{
  \item{x}{ an object of class "binMto", as can be calculated using \code{\link{binMto}} }
  \item{digits}{digits for rounding the output}
  \item{\dots}{further arguments to be passed to \code{print}}
}

\author{ Frank Schaarschmidt }

\keyword{ print }